+++ spec/display_manager_spec.rb
@@ -1,27 +1,20 @@
 require "spec_helper"
 
-RSpec.describe ArduinoCI::ArduinoCmd do
-  it "Finds the Arduino executable" do
-    arduino_cmd = ArduinoCI::ArduinoCmd.autolocate
-    # expect(arduino_cmd.path).not_to be nil
-  end
-end
-
-RSpec.describe ArduinoCI::ArduinoCmd::DisplayMgr do
+RSpec.describe ArduinoCI::DisplayManager do
   context "singleton ::instance" do
     it "produces an instance" do
-      expect(ArduinoCI::ArduinoCmd::DisplayMgr::instance).not_to be_nil
+      expect(ArduinoCI::DisplayManager::instance).not_to be_nil
     end
   end
 
   context "with_display" do
     it "Properly enables and disables" do
-      manager = ArduinoCI::ArduinoCmd::DisplayMgr::instance
+      manager = ArduinoCI::DisplayManager::instance
       expect(manager.enabled).to be false
       manager.with_display do |environment|
         expect(manager.enabled).to be true
         expect(environment.class).to eq(Hash)
-        also_manager = ArduinoCI::ArduinoCmd::DisplayMgr::instance
+        also_manager = ArduinoCI::DisplayManager::instance
         expect(also_manager.enabled).to be true
       end
       expect(manager.enabled).to be false